+++ test/mockFetch.ts
@@ -1,2 +1,0 @@
-import { enableFetchMocks } from 'jest-fetch-mock';
-enableFetchMocks();